RDA Germany 2020 poster
Towards FAIR Tropospheric Ozone assessment Report (TOAR) data center and Artificial Intelligence for Air Quality (IntelliAQ) data services and infrastructures

The IntelliAQ is a project aiming to develop novel approaches for the analysis and synthesis of global air quality data based on deep neural networks. The IntelliAQ project builds on the linkage among several different datasets that are collected from various repositories around the world. Its primary data source is the collection of the world's global air quality data from almost 10,000 measurement sites around in the context of IGAC's Tropospheric Ozone Assessment Report (TOAR). As of November 2019, the second phase (TOAR-II) that will provide updated and extended information on tropospheric ozone. This database will be complemented with data from the world's leading effort to collect global air pollutant measurements in near real-time and combined with high-resolution geodata, weather information, and satellite retrievals of atmospheric composition, e.g., the ERA5 reanalysis of the European Centre for Medium-range Weather Forecast. In the course of this effort, we established the TOAR – Datacenter as the unit that ran the TOAR database and maintained the TOAR-data portal. Through IntelliAQ, some of these datasets will be made available in aggregated form, primarily as web services with an interactive user interface (JOIN), a REST API to allow machine access (cf. JOIN). The project will also generate new datasets obtained through the fusion of the existing data and analysis by novel machine learning methods. In collaboration with EUDAT services primarily B2SHARE, we provide a platform for individual contributors to publish their ozone database as a companion with TOAR databases as a new EUDAT community with DOI (Digital Object Identifier). TOAR data center and IntelliAQ aims to make all of its products freely and openly accessible. We are trying to make our data findable and accessible by adopting the ISO 19115, INSPIRE, and WIGOS metadata standard and contributing to the development of the metadata standards in ENVRI-FAIR and the upcoming German national research data infrastructure (NFDI). The metadata profile is being developed as a community extension of the B2SHARE core metadata and will be integrated into EUDAT's search catalog B2FIND from where it can be harvested. To keep our data interoperable and reusable by using the community standard format such as simple ASCII interchange format or NetCDF4.0.
Moreover, web services follow standard APIs where these exist, and through B2FIND, the metadata can be automatically harvested and processed. IntelliAQ and TOAR aim to produce datasets that can be reused for several decades. Moreover, the TOAR data center furthering this effort by seeking a certification from Core-Trust seals to ensure that the TOAR data center will remain trustworthy, useful, and meaningful for the researcher in the long-term.